OIR-23: move the WC_Oxipay_Gateway into it's own class
